% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cramer}
\alias{cramer}
\title{Association between categorical variables}
\usage{
cramer(x)
}
\arguments{
\item{x}{data.frame or matrix.}
}
\value{
A matrix with the Cramer's V between the categorical variables.
}
\description{
Computes the Cramer's V, a measure of association between categorical
variables
}
\details{
Any integer variable is internally converted to a factor.
}
\note{
The original code was available at \url{http://sas-and-r.blogspot.nl/},
Example 8.39: calculating Cramer's V, posted by Ken Kleinman on Friday, June
3, 2011. As such, Ken Kleinman <\email{Ken_Kleinman@hms.harvard.edu}> is
entitled a \sQuote{contributor} to the R-package \pkg{pedometrics}.
}
\examples{
helpdata <- read.csv("http://www.math.smith.edu/r/data/help.csv")
data <- helpdata[, c("female", "homeless", "racegrp")]
cramer(data)
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
\references{
Cramér, H. Mathematical methods of statistics. Princeton: Princeton
University Press, p. 575, 1946.

Everitt, B. S. The Cambridge dictionary of statistics. Cambridge: Cambridge
University Press, p. 432, 2006.
}
\seealso{
\code{\link[vcd]{assocstats}}
}

