% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/numint.R
\name{is.numint}
\alias{is.numint}
\title{Numeric integer vectors}
\usage{
is.numint(x)
}
\arguments{
\item{x}{Object to be tested.}
}
\value{
\code{TRUE} or \code{FALSE} depending on whether \code{x} is a numeric
integer or not.
}
\description{
Test for objects of type \code{"numeric"} and \code{"integer"}.
}
\examples{
x <- 1:10
is.numint(x)
x <- as.numeric(x)
is.numint(x)
x <- c(1.1, 1, 1, 1, 2)
is.numint(x)
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
\seealso{
\code{\link[base]{is.numeric}}, \code{\link[base]{is.integer}}
}

