\name{pedometrics-package}
\alias{pedometrics-package}
\alias{pedometrics}
\docType{package}
\title{Pedometric Tools and Techniques}
\description{
This package contains many tools and techniques used in the field of pedometrics. These tools and techniques were developed as part of the PhD research project entitled Contribution to the Construction of Models for Predicting Soil Properties, developed by Alessandro Samuel-Rosa under the supervision of Dr. Lúcia Helena Cunha dos Anjos (Federal Rural University of Rio de Janeiro, Brazil), Dr. Gustavo de Mattos Vasques (Embrapa Soils, Brazil), and Dr. Gerard B M Heuvelink (ISRIC - World Soil Information, the Netherlands). Most functions are simple wrapper functions that extend the functionalities of other standard functions.
}
\details{
\tabular{ll}{
Package: \tab pedometrics\cr
Type: \tab Package\cr
Version: \tab 0.1-7\cr
Date: \tab 2014-06-16\cr
License: \tab GPL (>= 2)\cr
}
There are a lot of functions now.
}
\author{
Alessandro Samuel-Rosa

Maintainer: Alessandro Samuel-Rosa <alessandrosamuelrosa@gmail.com>
}
\references{
  There are many references now.
}
